fixed warnings

+++ bar/common/database.h
@@ -1379,7 +1379,7 @@ Errors Database_drop(const DatabaseSpecifier *databaseSpecifier,
                       );
 #else /* not NDEBUG */
   Errors __Database_open(const char              *__fileName__,
-                         ulong                   __lineNb__,
+                         size_t                  __lineNb__,
                          DatabaseHandle          *databaseHandle,
                          const DatabaseSpecifier *databaseSpecifier,
                          const char              *databaseName,
@@ -1401,7 +1401,7 @@ Errors Database_drop(const DatabaseSpecifier *databaseSpecifier,
   void Database_close(DatabaseHandle *databaseHandle);
 #else /* not NDEBUG */
   void __Database_close(const char     *__fileName__,
-                        ulong          __lineNb__,
+                        size_t         __lineNb__,
                         DatabaseHandle *databaseHandle
                        );
 #endif /* NDEBUG */
@@ -1632,7 +1632,7 @@ void Database_yield(DatabaseHandle *databaseHandle,
                     );
 #else /* not NDEBUG */
   bool __Database_lock(const char        *__fileName__,
-                       ulong             __lineNb__,
+                       size_t            __lineNb__,
                        DatabaseHandle    *databaseHandle,
                        DatabaseLockTypes lockType,
                        long              timeout
@@ -1654,7 +1654,7 @@ void Database_yield(DatabaseHandle *databaseHandle,
                       );
 #else /* not NDEBUG */
   void __Database_unlock(const char        *__fileName__,
-                         ulong             __lineNb__,
+                         size_t            __lineNb__,
                          DatabaseHandle    *databaseHandle,
                          DatabaseLockTypes lockType
                         );
@@ -2056,7 +2056,7 @@ Errors Database_removeColumn(DatabaseHandle *databaseHandle,
                                   );
 #else /* not NDEBUG */
   Errors __Database_beginTransaction(const char               *__fileName__,
-                                     uint                     __lineNb__,
+                                     size_t                   __lineNb__,
                                      DatabaseHandle           *databaseHandle,
                                      DatabaseTransactionTypes databaseTransactionType,
                                      long                     timeout
@@ -2094,7 +2094,7 @@ Errors Database_removeColumn(DatabaseHandle *databaseHandle,
   Errors Database_rollbackTransaction(DatabaseHandle *databaseHandle);
 #else /* not NDEBUG */
   Errors __Database_rollbackTransaction(const char     *__fileName__,
-                                        uint           __lineNb__,
+                                        size_t         __lineNb__,
                                         DatabaseHandle *databaseHandle
                                        );
 #endif /* NDEBUG */
@@ -2475,7 +2475,7 @@ Errors Database_deleteByIds(DatabaseHandle   *databaseHandle,
                         );
 #else /* not NDEBUG */
   Errors __Database_select(const char              *__fileName__,
-                           ulong                   __lineNb__,
+                           size_t                  __lineNb__,
                            DatabaseStatementHandle *databaseStatementHandle,
                            DatabaseHandle          *databaseHandle,
 // TODO: use DatabaseTable
@@ -2519,7 +2519,7 @@ bool Database_getNextRow(DatabaseStatementHandle *databaseStatementHandle,
   void Database_finalize(DatabaseStatementHandle *databaseStatementHandle);
 #else /* not NDEBUG */
   void __Database_finalize(const char        *__fileName__,
-                           ulong             __lineNb__,
+                           size_t            __lineNb__,
                            DatabaseStatementHandle *databaseStatementHandle
                           );
 #endif /* NDEBUG */
@@ -3080,7 +3080,7 @@ void Database_debugPrintLockInfo(const DatabaseHandle *databaseHandle);
 #ifdef NDEBUG
 void Database_debugPrintQueryInfo(const DatabaseStatementHandle *databaseStatementHandle);
 #else /* not NDEBUG */
-void __Database_debugPrintQueryInfo(const char *__fileName__, ulong __lineNb__, const DatabaseStatementHandle *databaseStatementHandle);
+void __Database_debugPrintQueryInfo(const char *__fileName__, size_t __lineNb__, const DatabaseStatementHandle *databaseStatementHandle);
 #endif /* NDEBUG */
 
 /***********************************************************************\
